replacing  PXAddMethodView for SplitableCardView
